pyproject.toml
[project]
name = "differential-equation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cupy-cuda12x==13.6.*",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pyqt5>=5.15.11",
]
